omnialpha/requirements.txt
httpx>=0.27.0
pandas>=2.2.0
pyarrow>=16.0.0
sqlalchemy>=2.0.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
streamlit>=1.36.0
plotly>=5.22.0
pytz>=2024.1
cryptography>=42.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
anthropic>=0.34.0
